feat: change block enttity `width` to Float

+++ src/main/java/com/mito/sectask/entities/Block.java
@@ -60,5 +60,5 @@ public class Block {
     private Block next;
 
     @Column(name = "width", nullable = true)
-    private Integer width;
+    private Float width;
 }
